move restart tests back to tn binary due to test setup calling the bin locally - correct another circular dependency
