+++ bin/telcoin-network/Cargo.toml
@@ -41,8 +41,20 @@ rayon = { workspace = true }
 [dev-dependencies]
 tempfile = { workspace = true }
 
+# restart tests
+telcoin-network = { path = ".", features = ["test-utils"] }
+ethereum-tx-sign = "6.1.3"
+nix = { version = "0.29", features = ["signal"] }
+serde_json = { workspace = true }
+tokio = { workspace = true }
+tracing-subscriber = { workspace = true }
+const-hex = "1.12.0"
+# Alternate sync json rpc lib- just for integration testing.
+jsonrpc = "0.18.0"
+
 [features]
 default = []
+test-utils = []
 faucet = ["tn-faucet"]
 
 [build-dependencies]
