+++ bin/telcoin-network/tests/it/main.rs
@@ -1,0 +1,5 @@
+//! IT test for main binary.
+
+mod restarts;
+
+async fn main() {}
